Добавляет деструктуризацию page-main
