+++ .vs/1618231-six-cities-15/v17/DocumentLayout.json
@@ -3,12 +3,16 @@
   "WorkspaceRootPath": "D:\\react\\1618231-six-cities-15\\",
   "Documents": [
     {
-      "AbsoluteMoniker": "D:0:0:{A2FE74E1-B743-11D0-AE1A-00A0C90FFFC3}|\u003CMiscFiles\u003E|D:\\react\\1618231-six-cities-15\\tsconfig.json||{90A6B3A7-C1A3-4009-A288-E2FF89E96FA0}",
-      "RelativeMoniker": "D:0:0:{A2FE74E1-B743-11D0-AE1A-00A0C90FFFC3}|\u003CMiscFiles\u003E|solutionrelative:tsconfig.json||{90A6B3A7-C1A3-4009-A288-E2FF89E96FA0}"
+      "AbsoluteMoniker": "D:0:0:{A2FE74E1-B743-11D0-AE1A-00A0C90FFFC3}|\u003CMiscFiles\u003E|D:\\react\\1618231-six-cities-15\\src\\entities\\places-list\\ui\\places-list.tsx||{0F2454B1-A556-402D-A7D0-1FDE7F99DEE0}",
+      "RelativeMoniker": "D:0:0:{A2FE74E1-B743-11D0-AE1A-00A0C90FFFC3}|\u003CMiscFiles\u003E|solutionrelative:src\\entities\\places-list\\ui\\places-list.tsx||{0F2454B1-A556-402D-A7D0-1FDE7F99DEE0}"
     },
     {
       "AbsoluteMoniker": "D:0:0:{A2FE74E1-B743-11D0-AE1A-00A0C90FFFC3}|\u003CMiscFiles\u003E|D:\\react\\1618231-six-cities-15\\src\\pages\\offer\\ui\\offer.tsx||{0F2454B1-A556-402D-A7D0-1FDE7F99DEE0}",
       "RelativeMoniker": "D:0:0:{A2FE74E1-B743-11D0-AE1A-00A0C90FFFC3}|\u003CMiscFiles\u003E|solutionrelative:src\\pages\\offer\\ui\\offer.tsx||{0F2454B1-A556-402D-A7D0-1FDE7F99DEE0}"
+    },
+    {
+      "AbsoluteMoniker": "D:0:0:{A2FE74E1-B743-11D0-AE1A-00A0C90FFFC3}|\u003CMiscFiles\u003E|D:\\react\\1618231-six-cities-15\\src\\pages\\page-main\\ui\\page-main.tsx||{0F2454B1-A556-402D-A7D0-1FDE7F99DEE0}",
+      "RelativeMoniker": "D:0:0:{A2FE74E1-B743-11D0-AE1A-00A0C90FFFC3}|\u003CMiscFiles\u003E|solutionrelative:src\\pages\\page-main\\ui\\page-main.tsx||{0F2454B1-A556-402D-A7D0-1FDE7F99DEE0}"
     }
   ],
   "DocumentGroupContainers": [
@@ -18,23 +22,44 @@
       "DocumentGroups": [
         {
           "DockedWidth": 200,
-          "SelectedChildIndex": 1,
+          "SelectedChildIndex": 3,
           "Children": [
             {
               "$type": "Bookmark",
               "Name": "ST:0:0:{3ae79031-e1bc-11d0-8f78-00a0c9110057}"
             },
             {
+              "$type": "Bookmark",
+              "Name": "ST:128:0:{116d2292-e37d-41cd-a077-ebacac4c8cc4}"
+            },
+            {
+              "$type": "Bookmark",
+              "Name": "ST:129:0:{116d2292-e37d-41cd-a077-ebacac4c8cc4}"
+            },
+            {
               "$type": "Document",
               "DocumentIndex": 0,
-              "Title": "tsconfig.json",
-              "DocumentMoniker": "D:\\react\\1618231-six-cities-15\\tsconfig.json",
-              "RelativeDocumentMoniker": "tsconfig.json",
-              "ToolTip": "D:\\react\\1618231-six-cities-15\\tsconfig.json",
-              "RelativeToolTip": "tsconfig.json",
-              "ViewState": "AQIAAAAAAAAAAAAAAAAAAAAAAAAAAAAA",
-              "Icon": "ae27a6b0-e345-4288-96df-5eaf394ee369.001642|",
-              "WhenOpened": "2024-02-16T19:09:18.362Z",
+              "Title": "places-list.tsx",
+              "DocumentMoniker": "D:\\react\\1618231-six-cities-15\\src\\entities\\places-list\\ui\\places-list.tsx",
+              "RelativeDocumentMoniker": "src\\entities\\places-list\\ui\\places-list.tsx",
+              "ToolTip": "D:\\react\\1618231-six-cities-15\\src\\entities\\places-list\\ui\\places-list.tsx",
+              "RelativeToolTip": "src\\entities\\places-list\\ui\\places-list.tsx",
+              "ViewState": "AQIAAAAAAAAAAAAAAAAAAAoAAAABAAAA",
+              "Icon": "ae27a6b0-e345-4288-96df-5eaf394ee369.003213|",
+              "WhenOpened": "2024-02-17T05:26:28.93Z",
+              "EditorCaption": ""
+            },
+            {
+              "$type": "Document",
+              "DocumentIndex": 2,
+              "Title": "page-main.tsx",
+              "DocumentMoniker": "D:\\react\\1618231-six-cities-15\\src\\pages\\page-main\\ui\\page-main.tsx",
+              "RelativeDocumentMoniker": "src\\pages\\page-main\\ui\\page-main.tsx",
+              "ToolTip": "D:\\react\\1618231-six-cities-15\\src\\pages\\page-main\\ui\\page-main.tsx",
+              "RelativeToolTip": "src\\pages\\page-main\\ui\\page-main.tsx",
+              "ViewState": "AQIAAEEAAAAAAAAAAAAAAF0AAAAAAAAA",
+              "Icon": "ae27a6b0-e345-4288-96df-5eaf394ee369.003213|",
+              "WhenOpened": "2024-02-17T05:22:05.186Z",
               "EditorCaption": ""
             },
             {
@@ -45,14 +70,10 @@
               "RelativeDocumentMoniker": "src\\pages\\offer\\ui\\offer.tsx",
               "ToolTip": "D:\\react\\1618231-six-cities-15\\src\\pages\\offer\\ui\\offer.tsx",
               "RelativeToolTip": "src\\pages\\offer\\ui\\offer.tsx",
-              "ViewState": "AQIAAAAAAAAAAAAAAAAAAAAAAAAFAAAA",
+              "ViewState": "AQIAAOsAAAAAAAAAAAArwPwAAAAjAAAA",
               "Icon": "ae27a6b0-e345-4288-96df-5eaf394ee369.003213|",
               "WhenOpened": "2024-02-16T19:08:15.772Z",
               "EditorCaption": ""
-            },
-            {
-              "$type": "Bookmark",
-              "Name": "ST:0:0:{cce594b6-0c39-4442-ba28-10c64ac7e89f}"
             }
           ]
         }
