+++ src/features/plasces-sorting/index.ts
@@ -1,0 +1,3 @@
+import PlacesSorting from './ui/places-sorting';
+
+export {PlacesSorting};
