--- src/shared/AppRoutes.ts
@@ -1,7 +1,0 @@
-export enum AppRoutes {
-    Main = '/',
-    Offer = '/offer',
-    Favorites = '/favorites',
-    Login = '/login'
-}
-
