--- src/shared/AutorizationStatus.ts
@@ -1,5 +1,0 @@
-export enum AutorizationStatus {
-    Auth = 'AUTH',
-    NoAuth = 'NO_AUTH',
-    Unkhown = 'ANKHNOWN'
-}
