+++ src/shared/constans.ts
@@ -1,0 +1,13 @@
+export enum AutorizationStatus {
+  Auth = 'AUTH',
+  NoAuth = 'NO_AUTH',
+  Unkhown = 'ANKHNOWN'
+}
+
+export enum AppRoutes {
+  Main = '/',
+  Offer = '/offer',
+  Favorites = '/favorites',
+  Login = '/login'
+}
+
