+++ src/widgest/header/index.ts
@@ -1,0 +1,3 @@
+import Header from './ui/header';
+
+export {Header};
