+++ src/widgest/locations/index.ts
@@ -1,0 +1,3 @@
+import Locations from './ui/locations';
+
+export {Locations};
