+++ src/widgest/map/index.ts
@@ -1,0 +1,3 @@
+import Map from './ui/map';
+
+export {Map};
