+++ src/widgest/map/ui/map.tsx
@@ -1,0 +1,6 @@
+
+const Map = () => (
+  <section className="cities__map map"></section>
+);
+
+export default Map
