+++ src/widgest/place-card/ui/place-card.tsx
