+++ src/widgest/places-list/index.ts
@@ -1,0 +1,3 @@
+import PlacesList from './ui/places-list';
+
+export { PlacesList };
