+++ src/widgest/places-list/ui/places-list.tsx
@@ -1,0 +1,13 @@
+import { PlaceCard } from 'src/widgest/place-card';
+
+
+const PlacesList = () => (
+  <div className="cities__places-list places__list tabs__content">
+    <PlaceCard />
+    <PlaceCard />
+    <PlaceCard />
+    <PlaceCard />
+  </div>
+)
+
+export default PlacesList;
